(FM-2040) (FM-2042) Fix pool members
So it turns out trying to manage the pool members from the f5_pool
resource doesn't make sense, since all of the parameters except `port`
can be configured through the f5_node resource.

So now the `members` hash only accepts `name` (which must be the name of
an existing f5_node) and `port` (which is 0-65535).