Only display f5_node non-default route domain
BigIP nodes have addresses that may end with %<route domain id> which is
an integer. From the API perspective, if the node's route domain id is 0
then it does not display it, otherwise it does.

From the GUI perspective, if the node's route domain id is the default
then it does not display it, otherwise it does.

This checks the partitions default route domain ID, and only prints it
if the node route domain is non-default.