(maint) Use current Beaker and do not override ssh config
Previously we pinned to an old version of Beaker because the API for the
EC2 hypervisor changed.  Using an old version of Beaker reduced our
confidence in debugging and merging our code back into Beaker is a
stated goal.  Since that's the case I updated us to the current Beaker
and updated audited the custom Hypervisor code so that we can be more
confident in debugging and are closer to being able to merge our work
into Beaker.

The initial impetus for this was Beaker being unable to log in once we
had overridden the host['ssh'] hash with our computed password (it seems
like the behavior may have changed between 2.10 and 2.18.  To solve this
issue we updated `enable_root_f5` (previously `enable_root`) to
non-destructively set the password value _after_ having successfully
changed the remote password.