Fix type requires
When the master autoloads the type during attribute validation, the
module's lib/ directory is not in the ruby libpath and thus any RUBYLIB
requires in types that are not autoloaded will fail. To fix this, the
requires can be given absolute paths to the libraries needed. This will
work for pluginsync'd libraries or autoloaded types.